backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.9
requests==2.31.0
pypdf==3.17.0
python-multipart==0.0.6
httpx==0.24.1